bugfix: don't leak sockets when handling redirects, and honor provided http agent
When handling a 302 redirect response with keepAlive option, the response body
must be fully read to allow the socket to be put in the 'free socket' bucket of
the http agent.  Otherwise it stays open for ever.

Additionally, the agent provided in the user options was not transmitted to the
http/https request() call.